[Q&A] MAX_CID is a positive integer

+++ app/performance/test_performance.c
@@ -116,7 +116,7 @@ static int test_compression_perfs(const bool is_verbose,
                                   char *filename,
                                   const rohc_cid_type_t cid_type,
                                   const size_t wlsb_width,
-                                  const unsigned int max_contexts,
+                                  const size_t max_contexts,
                                   unsigned long *packet_count);
 static int time_compress_packet(struct rohc_comp *comp,
                                 unsigned long num_packet,
@@ -127,7 +127,7 @@ static int time_compress_packet(struct rohc_comp *comp,
 static int test_decompression_perfs(const bool is_verbose,
                                     char *filename,
                                     const rohc_cid_type_t cid_type,
-                                    const unsigned int max_contexts,
+                                    const size_t max_contexts,
                                     unsigned long *packet_count);
 static int time_decompress_packet(struct rohc_decomp *decomp,
                                   unsigned long num_packet,
@@ -263,7 +263,7 @@ int main(int argc, char *argv[])
 		cid_type = ROHC_SMALL_CID;
 
 		/* the maximum number of ROHC contexts should be valid */
-		if(max_contexts < 1 || max_contexts > (ROHC_SMALL_CID_MAX + 1))
+		if(max_contexts < 1 || (size_t) max_contexts > (ROHC_SMALL_CID_MAX + 1))
 		{
 			fprintf(stderr, "the maximum number of ROHC contexts should be "
 			        "between 1 and %u\n\n", ROHC_SMALL_CID_MAX + 1);
@@ -276,7 +276,7 @@ int main(int argc, char *argv[])
 		cid_type = ROHC_LARGE_CID;
 
 		/* the maximum number of ROHC contexts should be valid */
-		if(max_contexts < 1 || max_contexts > (ROHC_LARGE_CID_MAX + 1))
+		if(max_contexts < 1 || (size_t) max_contexts > (ROHC_LARGE_CID_MAX + 1))
 		{
 			fprintf(stderr, "the maximum number of ROHC contexts should be "
 			        "between 1 and %u\n\n", ROHC_LARGE_CID_MAX + 1);
@@ -382,7 +382,7 @@ static int test_compression_perfs(const bool is_verbose,
                                   char *filename,
                                   const rohc_cid_type_t cid_type,
                                   const size_t wlsb_width,
-                                  const unsigned int max_contexts,
+                                  const size_t max_contexts,
                                   unsigned long *packet_count)
 {
 	pcap_t *handle;
@@ -617,7 +617,7 @@ error:
 static int test_decompression_perfs(const bool is_verbose,
                                     char *filename,
                                     const rohc_cid_type_t cid_type,
-                                    const unsigned int max_contexts,
+                                    const size_t max_contexts,
                                     unsigned long *packet_count)
 {
 	const struct rohc_ts arrival_time = { .sec = 0, .nsec = 0 };
